Fix confusion between path, sequence, and pattern
The word pattern is no longer used.
A set of paths is given, and common patterns of nodes are referred to as sequences.

It might be that in future we decide it is better to use pattern instead of sequence. Sequence was used however to emphasise that a pattern is linear.