Revert to old pathContainsPattern implementation
For reasons, the new implementation only worked
in test cases and not for the smoke test. It may
be that this is related to the bug mentioned
before that structural equality is not recursive,
meaning it does not check expressions within
expressions. This means that it works fine for
mocks but not complex statements.

Due to there being multiple possible causes
there is no way of saying this with absolute
certainty, but it would perhaps be possible to
retry this implementation once the new recursive
structural equality is implemented.