Add out keyword to generics for project types
This allows one to use a super type of a generic in a class.

For instance, one might mine JavaMavenProjects first. Afterwards
however, one may to wish to generate a graphs, in which case it is only
relevant that it is a JavaProject. As long as the liskov substitution
principle as adhered to it should be permissible to use the parent
interfaces.

This can be seen in the LibraryUsageGraphGenerator class. This class
takes JavaProjects and outputs JavaProjects. However it gets a
JavaMavenProject as input for the library.