Wrap mined projects in JavaMavenProject
This reflects the use case where mined projects are source code
instead of jars. This will often be the case on GitHub.

Note that it will still be possible to mine JavaJarProjects in the
future, as the pipeline can also handle these. It may be relevant in
future to make several presets of pipelines, or perhaps think about ways
of dynamically constructing pipelines.