Fix Detekt rule violations
No idea how they got into `master`, but they did, so here are my fixes for them. I agree with the rule preferring 'to' expressions over 'Pair' statements - feel free to disagree, however.

I did not agree with the 'MandatoryBracesIfStatements' rule. Seeing as we don't consistently follow it in the rest of the codebase, either, I turned it off in `detekt.yml`.