Make copy of all nodes in a path
The PathEnumerator previously reused node instances across paths. This led to the possibility that two paths having overlap of nodes shared the same node instances, and thus also the same Jimple statement instances. This caused unwanted behavior when modifying either of the copies (e.g. in the removal of invalid targets).