Remove copy-dependencies step in Maven compiler
This turned out not to be necessary for the pipeline (analyzing the compiled project for LUGs still works without it), and was quite time and space intensive. The smoke test has been modified to include an external dependency in one of the user projects, do demonstrate that the mining pipeline still works.