Add progress bar with progress unknown
This is less misleading than having a progress bar which depends on
an iterable which is not indicative of the actual progress.
Now an empty progress bar is shown which does still gives the time
elapsed, and at the end sets the maximum to the count of the produced
iterator.