Add mining flavor argument to CMD instructions
The program is actually expecting a first argument (the directory flavor). Since this is not handled by our CMD library, it doesn't show up in the generated argument help menu. Still, we should add it to our README instructions, because I had to manually dive into the code to discover this (which normal users shouldn't have todo).