Replace Invalid Targets in actual method, not duplicated statement list
Before, we would replace invalid targets on the list of duplicate
statements, which was a duplicate of the active body and was thrown
away afterwards.

The issue was that there was no guarantee that this list of statements
ended with a return statement. Therefore, if the last statement was an
`if` statement with an invalid target, we replaced its body with a
reference to the last statement. This would be the `if` statement itself
creating an empty while loop, instead of the intended behaviour which
was a reference to the `return` or `returnvoid` of the method body.

Now, we do the same but on the active body on the statement. This is
guaranteed to have a return statement as the method addReturnStatement
was called before.