Fix issue with EvoSuite JAVA_HOME setup
On my macOS setup, the smoke test failed, due to EvoSuite not finding 'tools.jar'. After some debugging, I found out that this had to do with the JAVA_HOME variable not being visible to EvoSuite. I have fixed this by setting this variable explicitly for the subprocess that we launch for EvoSuite.