Turn off EvoSuite VFS usage
Currently, EvoSuite used a VFS for mocking IO-related functionality in its tests. However, due to a discrepancy between the runtime behavior of EvoSuite tests and their search-time environment, this caused tests to fail when executed. This PR adds two options which disable the usage of virtualized file systems and network.